Fixed modules. Added prebody field. Added spinners to modules.
Problem was that no check was done on the arrays to see if they were emtpy or not. Placed an if statement around it.  The probody field is added so that actions can be performed on modules that require data before this data is received. This only works for modules that have customContainer set to true.