tooltips now higher z-index than menu
Wide tooltips could go under the menu bar on the left. Now they have a z-index of 999. The initial display state of the tooltips were initial, allowing them to be seen in the top left corner when not used, so changed that into none in the lib file to hide it from the start.